create component: PhotoItem component
