+++ src/app/components/photoList/PhotoList.tsx
@@ -1,14 +1,9 @@
-import { Photo, PhotoForList, PhotoResponse } from '@/types/unsplash';
+import { PhotoForList } from '@/types/unsplash';
 import styles from './photoList.module.css'
-import Image from "next/image";
 import LoadingSpinner from '../loadingSpinner/LoadingSpinner';
 import Pagination from '../pagination/Pagination';
-import Modal from '../modal/Modal';
-import useLocalStorage from '@/app/hooks/useLocalStorage';
 import { useState } from 'react';
-import bookmarkWhite from "/public/images/bookmark_white.svg";
-import bookmarkFill from "/public/images/bookmark_fill.svg";
-import { fetchUnsplashPhotoBy } from '@/app/services/photo.service';
+import PhotoItem from '../photoItem/PhotoItem';
 
 type PhotoListProps = {
     isLoading: boolean;
@@ -23,51 +18,9 @@ type PhotoListProps = {
 const PhotoList: React.FC<PhotoListProps> = ({ isLoading, photos, photoListCurrentPage, searchTotal, searchTerm, onPageChange, onBookmarkChange }) => {
     const [currentPage, setCurrentPage] = useState(photoListCurrentPage);
 
-    // NOTE(hajae): 로그인 기능이 따로 구현이 되어있지 않기 때문에 로컬스토리지에 북마크를 저장/관리
-    const [bookmarks, setBookmarks] = useLocalStorage<PhotoForList[]>('bookmarks', []);
-
-    const [selectedPhoto, setSelectedPhoto] = useState<Photo | null>(null);
-    const [isModalOpen, setIsModalOpen] = useState(false);
-
-    const fetchPhotoById = (photoId: string) => {
-        fetchUnsplashPhotoBy(photoId)
-            .then((res: PhotoResponse) => { setSelectedPhoto(res.photos); });
-    }
-
-    const handleImageClick = (photo: PhotoForList) => {
-        setIsModalOpen(true);
-        fetchPhotoById(photo.id);
-    };
-
-    const isBookmarked = (photoId: string): boolean => {
-        return bookmarks.some((bookmark) => bookmark.id === photoId)
-    }
-
-    const closeModal = () => {
-        setIsModalOpen(false);
-        setSelectedPhoto(null);
-    };
-
     const handlePageChange = (page: number) => {
         setCurrentPage(page);
         onPageChange(page, searchTerm);
-    };
-
-    const handleBookmarkClick = (photo: PhotoForList) => {
-        const existingBookmarkIndex = bookmarks.findIndex((bookmark) => bookmark.id === photo.id);
-
-        // NOTE(hajae): 이미 북마크에 있는 경우 제거, 아닌 경우 추가
-        if (existingBookmarkIndex !== -1) {
-            const updatedBookmarks = [...bookmarks];
-            updatedBookmarks.splice(existingBookmarkIndex, 1);
-            localStorage.setItem('bookmarks', JSON.stringify(updatedBookmarks));
-            setBookmarks(updatedBookmarks);
-            if (onBookmarkChange) onBookmarkChange();
-        } else {
-            const updatedBookmarks = [...bookmarks, photo];
-            localStorage.setItem('bookmarks', JSON.stringify(updatedBookmarks));
-            setBookmarks(updatedBookmarks);
-        }
     };
 
     // NOTE(hajae): 사진이 로딩, 유무에 따른 표시
@@ -84,22 +37,11 @@ const PhotoList: React.FC<PhotoListProps> = ({ isLoading, photos, photoListCurre
             )
         } else {
             return photos.map((photo: PhotoForList) => (
-                <div className={styles.PhotoImageContainer} key={photo.id}>
-                    <Image
-                        className={styles.PhotoImage}
-                        src={photo.urls.small}
-                        alt={photo.alt_description}
-                        width={300}
-                        height={300}
-                        onClick={() => handleImageClick(photo)}
-                    />
-                    <Image
-                        className={styles.Bookmark}
-                        src={isBookmarked(photo.id) ? bookmarkFill.src : bookmarkWhite.src}
-                        alt="bookmark"
-                        width={30}
-                        height={30}
-                        onClick={() => handleBookmarkClick(photo)} />
+                <div className={styles.PhotoImageContainer}>
+                    <PhotoItem 
+                        key={photo.id}
+                        photo={photo}
+                        onBookmarkChange={onBookmarkChange}/>
                 </div>
             ));
         }
@@ -113,9 +55,6 @@ const PhotoList: React.FC<PhotoListProps> = ({ isLoading, photos, photoListCurre
             {!isLoading && photos.length !== 0 && <div>
                 <Pagination currentPage={currentPage} totalItems={searchTotal} onPageChange={handlePageChange} />
             </div>}
-            {isModalOpen && selectedPhoto && (
-                <Modal onClose={closeModal} photoInfo={selectedPhoto} onBookmarkClick={handleBookmarkClick} />
-            )}
         </div>
     )
 }
